+++ baybe/constraints/continuous.py
@@ -81,6 +81,8 @@ class ContinuousLinearConstraint(ContinuousConstraint):
                 "The given 'coefficients' list must have one floating point entry for "
                 "each entry in 'parameters'."
             )
+        if any(c == 0.0 for c in coefficients):
+            raise ValueError("All entries in 'coefficients' must be non-zero.")
 
     @coefficients.default
     def _default_coefficients(self) -> tuple[float, ...]:
@@ -257,16 +259,16 @@ class ContinuousCardinalityConstraint(
             corresponding parameter names.
         """
         # The number of possible parameter configuration per set cardinality
-        n_configurations_per_cardinality = [
-            math.comb(len(self.parameters), n)
-            for n in range(self.min_cardinality, self.max_cardinality + 1)
-        ]
+        n_configs_per_cardinality = np.array(
+            [
+                math.comb(len(self.parameters), n)
+                for n in range(self.min_cardinality, self.max_cardinality + 1)
+            ]
+        )
 
         # Probability of each set cardinality under the assumption that all possible
         # inactive parameter sets are equally likely
-        probabilities = n_configurations_per_cardinality / np.sum(
-            n_configurations_per_cardinality
-        )
+        probabilities = n_configs_per_cardinality / n_configs_per_cardinality.sum()
 
         # Sample the number of active/inactive parameters
         n_active_params = np.random.choice(